BackEnd/src/main/java/com/example/BackEnd/Controllers/UserVerificationController.java
package com.example.BackEnd.Controllers;

import com.example.BackEnd.Services.UserVerification;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RequiredArgsConstructor
@RestController
@RequestMapping("/api/verification")
public class UserVerificationController {

    private final UserVerification userService;

    @GetMapping
    public ResponseEntity<String> verifyUser(@RequestParam("token") String token) {
        try {
            userService.verifyUser(token);
            return ResponseEntity.status(HttpStatus.OK).body("User verified successfully");
        } catch (UsernameNotFoundException e) { // user not found
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
        } catch (IllegalStateException e) { // user already verified
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
        }
    }
}
